+++ examples/springboot/src/main/java/com/inrupt/client/examples/springboot/service/IBookLibraryService.java
@@ -18,7 +18,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.inrupt.client.examples.springboot;
+package com.inrupt.client.examples.springboot.service;
 
 import com.inrupt.client.examples.springboot.model.Book;
